requirements.txt
Django>=4.2.2,<4.3
djangorestframework>=3.14,<3.15
psycopg2>=2.9.6,<2.10
drf-spectacular>=0.26.3,<0.27
Pillow>=10.0.0,<11.0.0